Move general helper function to reduction module.
